FS change. Cleaned things up. Added init system
