--- Notes/Multitasking.md
@@ -1,7 +1,0 @@
-Skynet Multitasking System
-
-
-
-
-Question: How taxing is it to save information to the turtle?
-Question: What happens when yielding during data.get during a task that yields? More generally, what happens when we have multiple yields?
