+++ src/etc/skynetrc
@@ -1,0 +1,3 @@
+/exec/hello.lua
+/exec/class_test.lua
+/exec/emulator_close.lua
