+++ src/exec/emulator_close.lua
@@ -1,0 +1,4 @@
+--Close the emulator if skynet is running on one.
+if os.exit then
+    os.exit(0)
+end
