+++ src/init/init.cfg
@@ -1,0 +1,9 @@
+{
+    LOG = {
+        LOG_DIR = '/sys/log',
+        DEFAULT_LOGS = {
+            MAIN = 'recent.log'
+        }
+    },
+    POST_INIT = '/lib/core/start.lua'
+}
