+++ src/init/nym/sugar.lua
@@ -1,0 +1,3 @@
+-- SOME HELPER STUFF --
+EMPTY_PROPERTY 	= '__empty__'
+EMPTY_BOOL 		= false
