--- src/lib/deprecated/task/Macro.lua
@@ -1,1 +1,0 @@
--- Macro tasks are collections of micro tasks that accomplish a particular objective.
