--- src/sys/skynetrc
@@ -1,2 +1,0 @@
-/MWP/sys/hello.lua
-/MWP/tests/taskSequence_test.lua
