--- src/tests/GitTest.lua
@@ -1,3 +1,0 @@
-for i =  1,  6
-	do turtle.forward() 
-end
